feat: Migrate query docs to v2 layout
