+++ app/routes/query.$version.docs.framework.$framework.examples._index.tsx
@@ -2,9 +2,5 @@ import { redirect } from '@remix-run/node'
 import type { LoaderFunctionArgs } from '@remix-run/node'
 
 export const loader = (context: LoaderFunctionArgs) => {
-  const { framework } = context.params
-
-  throw redirect(
-    context.request.url.replace('/examples/', `/examples/${framework}/basic`)
-  )
+  throw redirect(context.request.url.replace('/examples/', `/examples/basic`))
 }
